For log-fft feature, some different ways to normalize in time and log domains.
Possibly, normalization should occur before or after conversion to
log-frequency domain.  Final mean removal is for PCA reasons, however
pre- or post-processing normalization may be a good idea if we want to
make sure that classification occurs based on shape of spectrum, but
not amplitude.

On the other hand it may be a good idea to allow amplitude as part of
the classification, so we leave some optional normalization lines
commented-out until we have the final training data to play with.